requirements.txt
alabaster==0.7.12
Babel==2.10.3
certifi==2022.9.24
charset-normalizer==2.1.1
docutils==0.19
idna==3.4
imagesize==1.4.1
importlib-metadata==5.0.0
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.4
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.4
requests==2.28.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.2.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
urllib3==1.26.12
zipp==3.9.0
